% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_types.R
\name{conv_prepare}
\alias{conv_prepare}
\title{Preparation of Convolutional Covariate Representations}
\usage{
conv_prepare(
  object,
  W,
  dimyx,
  fractional,
  normalize,
  .direction.preferred = NULL
)
}
\arguments{
\item{normalize}{}
}
\description{
Preparation of Convolutional Covariate Representations
}
