% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_types.R
\name{Pcov}
\alias{Pcov}
\title{Pcov
Pcov is meant for point process valued data,
an object of type 'ppp' or sf POINT objects can be handled as data input
Pcov can be used to process multiple covariates at the same time, each covariate will be given the name of the argument if provided,
or otherwise will simply be assigned the name of the symbol passed in
Pcov returns a representation of the point process containing the necessary components for setting up a convolutional basis representation.
The return value is a dummy vector of length 1 containing only placeholder numeric data (this may be used later to store genuinely useful information).
The attributes contain angle orientation and distance matrices for setup of the convolutional basis. It also contains the precomputed fft of the
pixellated version of the process
additional attributes include the observation window of the process and the dimensions of the process
a number of vctrs methods need to still be added for Pcovs, such as restore methods etc}
\usage{
Pcov(..., W, dimyx, fractional = FALSE)
}
\description{
Pcov
Pcov is meant for point process valued data,
an object of type 'ppp' or sf POINT objects can be handled as data input
Pcov can be used to process multiple covariates at the same time, each covariate will be given the name of the argument if provided,
or otherwise will simply be assigned the name of the symbol passed in
Pcov returns a representation of the point process containing the necessary components for setting up a convolutional basis representation.
The return value is a dummy vector of length 1 containing only placeholder numeric data (this may be used later to store genuinely useful information).
The attributes contain angle orientation and distance matrices for setup of the convolutional basis. It also contains the precomputed fft of the
pixellated version of the process
additional attributes include the observation window of the process and the dimensions of the process
a number of vctrs methods need to still be added for Pcovs, such as restore methods etc
}
