% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_types.R
\name{placeholder_value}
\alias{placeholder_value}
\title{covariate placeholders should carry the relevant information regarding
the appropriate single level entity  information needed in model fitting (such as mgcv::gam)
cases include an age, sex, geo-coordinate (x,y), or possibly higher order coordinates, (x,y,t, w) for extra w}
\usage{
placeholder_value(value)
}
\description{
covariate placeholders should carry the relevant information regarding
the appropriate single level entity  information needed in model fitting (such as mgcv::gam)
cases include an age, sex, geo-coordinate (x,y), or possibly higher order coordinates, (x,y,t, w) for extra w
}
