% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_types.R
\name{Rcov}
\alias{Rcov}
\title{Raster Type Covariates
Rcov handles data passed in raster formats.
Currently, only 'im' type objects are handled
Rcov does minimal handling of the raster objects, its primary goal is for metadata management and allowing natural manipulation of raster valued covariates}
\usage{
Rcov(..., interpolator = c("bilin", "bin", "spatstat"))
}
\arguments{
\item{interpolator}{options include 'bin' which treats the raster as a piecewise constant surface given by the raster values,
and 'bilin' which implements a bilinear interpolation based on assigning the weights of the pixels to their upper left coordinate corners
alternatively, with the argument centroid = TRUE bilinear interpolation will be computed on the basis of centroids.}
}
\description{
Raster Type Covariates
Rcov handles data passed in raster formats.
Currently, only 'im' type objects are handled
Rcov does minimal handling of the raster objects, its primary goal is for metadata management and allowing natural manipulation of raster valued covariates
}
